vktrace: Don't put VkAllocationCallbacks in trace file.
Allocation callbacks don't make sense in the trace file
because the callback code isn't available for replay.
Instead, put NULL in the trace file so that replay uses the
default allocators.

This fixes a problem with tri demo which uses allocator
callbacks for create and destroy instance.  Replay would
crash when playing back destroy instance because it jumped
to an invalid location. (create instance hard-coded a NULL)

Note that space is still reserved, although unused, in the
trace buffer for VkAllocationCallbacks.  This should not
be a huge space issue, since the Vulkan calls that use these
callbacks are relatively infrequent.
This commit does make the buffer allocation and the use of NULL
more consistent across the API functions.