layers:  Suppress swapchain warning message for the SDK release
On certain hardware, layer was spewing the following warning message in
the smoke demo:

Swapchain(PERF): vkAcquireNextImageKHR() called when the application already
owns all presentable images in this swapchain except for the image currently
being displayed.  This call to vkAcquireNextImageKHR() cannot succeed unless
another thread calls the vkQueuePresentKHR() function in order to release
ownership of one of the presentable images of this swapchain.