vkreplay: Handle errors when objects fail to remap.
In the current situation (while trying to trace / replay Doom), we're
missing the creation of a VkImage object, and so subsequent calls to
m_objMapper.remap_images(..) are failing. Previously, when we failed to
remap an object, the code would wrongly return the original object, which
would continue to get passed into the driver, and likely cause a crash due
to accessing invalid memory. This change will return VK_NULL_HANDLE if an
object fails to be remapped, and adds some validation that the objects
returned from a remap_*() call are not VK_NULL_HANDLE, and will return as
needed. This avoids the crash in the driver and reports an appropriate
error to the console.

This change only protects against those errors that I encountered while
trying to trace & replay Doom. There are MANY more places where remap_*()
is called that need to have similar protection added.

This change does NOT fix the missing VkImage object issue, simply protects
against invalid remaps.

XCAP-309

Change-Id: I8a5e8740c084f407d4cc3b1efddd46146bd6b108