vktrace trim: recreate images in-order of the application to maintain mem requirements.
On AMD systems the memory requirements are affected by the order of image
creation, and so we need to recreate them in the same order on replay in
order to ensure that memory allocations and calls to vkBindImageMemory()
will be successful. This change keeps the necessary image-related calls in a list,
including vkCreateImage() and vkDestroyImage() so that we can output them
in the same order and ensure in-order creation. These calls can be made
from multiple threads, so a critical section is used to protect the list.

* Put ordered image creation code into a preprocessor macro to make it easier
to enable / disable for different testing purposes and to make it more clear
which parts of the code are needed for ordered image creation and which parts
are NOT needed.

XCAP-312

Change-Id: I2c27a781eea15facd600f70901f32192d105d1c5