vktrace trim: Transition images into the proper layout prior to the first frame.
There's a lot of information that needs to be stored in order to properly
issue an ImageMemoryBarrier on a CommandBuffer in the correct Queue. This
seems to remove the validation errors that were being generated when the
image layouts were not correct, but I still don't have any visual
confirmation (in the rendering) to suggest that everything is working as
expected.

XCAP-312

Change-Id: Iab8af044b4ff499a4ce9c6560386d947c2a8cc5d