via: Add new tools application
Create an application (called via) which will perform
some validation of the Vulkan setup of a machine and spit out
a resulting HTML file.  This file should be used when
submitting Vulkan bugs related to the loader, layers, or
samples.

Eventually, it would be nice to run this tool as part of the
LunarG Vulkan SDK install to validate that everything
is working properly after installation is complete.

To run it, simply run with no parameters and it will write
a resulting file to the current directory, called
'via.html'.  If, for whatever reason, it is unable
to write to the current directory, it will write to the
user's home folder.

Change-Id: I37ae7d103cd319e9c92a5de1fdeb65a39ddc6a71