vktrace trim: Refactored some of the snapshot code for easier reuse.
No functional changes, just restructuring.

One CommandPool and one CommandBuffer are created per Device so that the
trim code can use them as needed. Previously they were specific to within
the snapshot method, but are now static and within the namespace, so they
can be used elsewhere and there is an accessor that will create them only
as-needed. Note that since we only create one per device, it does mean
that we may need to submit & wait idle if the commandBuffer needs to be
used for more complex operations.

Added three new methods to: 1) create a staging buffer, 2) transition
images, and 3) transition buffers.

XCAP-312

Change-Id: I08beaa2020579394ab11c071f134781a28c7cc79