vktrace trim: Don't try to copy contents into UNDEFINED images.
Images that have the UNDEFINED layout means that the current contents will
be discarded when the application uses it next, as such there's no reason
for us to generate the calls to recreate its contents.

This avoids a validation error caused by us trying to transition the
images back into the UNDEFINED layout after setting the image contents.

XCAP-383

Change-Id: Ib2193a8518df7b09389a174a0e3ebd7c5a9143c0