vktrace trim: Fix reference to map/unmap memory packets.
ObjectInfo.Buffer was being referenced when it was supposed to be
ObjectInfo.Image

This fixes the SaschaWillems SkeletalAnimation replay error about not
being able to remap a buffer. The replay is still not yet animated because
it uses persistenly mapped buffers and the trim code isn't yet hooked into
the pageguard solution.

XCAP-382

Change-Id: I1ddde8966a4e72506fb2dbc0a2eeb8bfdfd05ef0