vktrace trim: Create Fences as SIGNALED as needed.
Instead of submitting a queue to signal the necessary fences, we'll
simply create them as SIGNALED.

This fixes an issue seen while testing SaschaWillems raytracing example
which has a fence that is created as SIGNALED. Our previous code was
causing errors because the we were trying to signal this fence again using
a vkQueueSubmit(). Creating the fence as SIGNALED is the easier overall
solution.

XCAP-312

Change-Id: If78d369ce802718b09ae2674c2ea27884a3e5d88