vkreplay: Fix memory clean up after remap error.
If one of the Vulkan objects could not be remapped, then the code which
cleans up this memory would attempt to free a pointer allocated by the
application, causing heap corruption. This change ensures that the
pointers are NULL until allocated by vkreplay, so that we can properly
clean up the memory.

I'm not yet sure why the vkImageView (in this case) could not be remapped.
It was noticed while working on trim support, so will be considered a
separate issue.

XCAP-312

Change-Id: I0f15597dcfd81986d84573fd89859df4d0998b5d