vktrace trim: VKTRACE_TRIM_FRAMES now supports 'start,numFrames' format.
Instead of setting the start and end frame numbers in a 'start-end'
format, the environment variable can now use a 'start,numFrames' format to
specify the first frame to capture and the number of frames. As an
example, the previous method to capture frames 2000 through 2500 would
have been '2000-2500' whereas it can now also be done as '2000,500'.

This change also moves the evaluation of the environment variable into
trim::initialize() so that these changes are encapsulated inside the trim
functionality rather than inside the vktrace layer.

XCAP-312

Change-Id: I3c42acbd4a04838689bd22841ffff39dd05f3db5