vktrace trim: Fix recreation of LEVEL_SECONDARY command buffers.
The VK_COMMAND_BUFFER_LEVEL_END_RANGE is equal to the last valid value,
which is VK_COMMAND_BUFFER_LEVEL_SECONDARY, I had previously thought it
was one higher than that value. This change corrects the for-loop to
properly include it.

This was a primary fix for the SaschaWillems multithreading sample.

XCAP-312

Change-Id: I585193bfa882553c317fdbdb31d0618cb53dc85a