vktrace trim: Clean up trim logic on each entrypoint.
There was inconsistency before about which packets were copied and which
were deleted, so now the rule is that any packet that needs to get saved
while NOT in the trim frames should be copied and the original packet
deleted. While IN the trim frames, the original packet will be added to
the recorded_packets list.

This does cause a lot more allocating / deleting of memory, but it's a
simple to define pattern.

XCAP-312

Change-Id: I7ef004879e810fa3d559e463dabef5e4d4b587c8