vktrace trim: Only generate vkBindImageMemory() as needed.
Previously this call was being included in the list of image calls that
get recorded and replayed unconditionally. This would cause errors
during replay because sometimes the referenced memory object would no
longer exist (particularly if the image & memory were only used for
staging objects and had a short lifespan). In order to avoid the errors
during replay, this call is now included with the Image ObjectInfo so that
it is only generated for images that still exist at the start of the trim
frames.

XCAP-312

Change-Id: Ie8d89f33f965436edb49f2ec04ecccd8181cdb8e