vktrace trim: Support image layout transitions from RenderPasses.
This change tracks the framebuffer attachments and Renderpass layout
transitions and applies them to stored state within the command buffer.
This stored command buffer state is then applied to the actual images when
the command buffer is queued (by vkQueueSubmit).

Likewise, buffer access masks that are changed by vkCmdPipelineBarrier()
are also stored with the commandbuffer and applied at QueueSubmit.

XCAP-312

Change-Id: I29245be80024c006832c05afb88b805fba261f41