vktrace trim: Don't remove RenderPasses when they are destroyed.
This needs to be investigated further. It appears that several RenderPass
objects are being referenced when creating graphics pipeline objects, but
are then being deleted even though that graphics pipeline is still being
used. I haven't been able to find a place in the spec that confirms or
denies this behavior, so we should look into more deeply.

For now, if we don't remove the RenderPasses when they are supposed to be
deleted, then we avoid this issue.

XCAP-424

XCAP-312

Change-Id: Icf74998b32d4b24106a8c02586f619a3be2e02af