via: Fixes for Linux
Needed to add several more paths for drivers on Linux.  Also, look
in LD_LIBRARY_PATH for any so files if that's set.  Finally, fixed
several paths which would fail for warning cases.

With these changes, we've run it on Fedora Core 24 and Ubuntu with
AMD, Nvidia, and Intel (Mesa).

Thanks to Lenny and JohnZ for catching bugs and testing.

Change-Id: Ibdf1794f686693701359cb12354fbf7a2fa80461