vktrace trim: Fix parameters to mdd().
The parameter should be the object on which the call is being made,
previously we were always passing in the device, which could have resulted
in the wrong entrypoint being called. This didn't make any noticeable
difference, but in talking to LunarG this was identified as a bug in our
behavior.

XCAP-312

Change-Id: I38e0caa281e309dd9d0976c6396a4ac020d75b22