vktrace trim: Add support for recreating old RenderPasses.
We encountered situations where RenderPasses were being created, used to
create a pipeline, at some point being deleted, and then new RenderPasses
were created which happened to have the same handle as the previous one.
When vkreplay tried to replay the resulting calls, incompatible RenderPasses
were being used to create the Pipeline and causing undefined behavior.

This change keeps track of all the 'versions' of a RenderPass handle which are
created and tracks which version the Graphics Pipelines are created with.
When we it comes time to generate the trimmed trace file, the older
versions of the RenderPasses will be created strictly for the sake of
being able to create the corresponding Graphics Pipelines. The temporary
RenderPass is then immediately deleted as it appears to not be needed
otherwise.

XCAP-312

Change-Id: I2b4b224cc57b2922d99131722baf64194889faef