vktrace: add checksums to detect unchanged PMB pages
Added page checksums to detect when a traced app writes to
a pmb page without really modifying the contents of the
page. If /proc/self/pagemaps indicates a page is dirty but
the checksum for the page has not changed, the page is not
marked changed, so it does not get written to the trace file.

This change reduces the size of Dota2 trace files on Linux by
about 30%.  PMB page checksums are not used on Windows, they
would be difficult to implement because PAGE_GUARD is one-shot.

Also changed a "Network Connection Reset" message that is
logged when a traced app exits from an error message to a
verbose message.

Change-Id: If96a626f9861b8e6df9eaaf49e3fc2ab9e9a882e