vktrace: check for 0 rval in read from socket
If rval from recv() is 0, it indicates that the socket may
have closed. Modified code to check for an rval of 0 from
recv and return an error indicating the socket has been
closed.

This fixes a problem when tracing Dota2 on Linux - the
vktrace command would not exit when Dota2 was closed.

Change-Id: I8085f6856c74b66a76e685e1a5ccbe6a28de8391