vktrace trim: API calls are now streamed out like a regular trace.
After the State Snapshot is created, all snapshot objects are now written to trace file immediately;
API call packets are streamed into the file in real-time;
object deletion packets are written when / if the trim is completed.

If the application crashes unexpectedly, API call packets will be written up
until the point of the crash, but the object deletion packets would not
get written.

XCAP-536

Change-Id: Ia959576f84b7bbba2cf983457699a18ddaec4c3f