screenshot: Fix screenshot layer + trim compatibility.
Screenshot layer was not handling the case that the most recently created
CommandPool object was also being destroyed, and thus was trying to use a
destroyed object. The layer now keeps a most-recently-created list of
CommandPools, and will utilize the one that was most recently created and
not yet destroyed. If there are no available CommandPools, then the
screenshot layer will not generate a file.

Fixes issue #182

Change-Id: I8f248b44272f6b414f6ef578b962a2584c200e62