vktrace: Fix creation of trace from a vkreplay
The command:
   vktrace -p vkreplay -a '-t c.vktrace'
now works.

vkreplay was creating and destroying a vkinstance so as
to query the gpu and driver version. This resulted in
the trace layer getting loaded twice, and thus two trace
file headers being sent to vktrace. Changed code
so that gpu and driver version is not queried on vkreplay
startup, but saved from result of the replay of
vkGetPhysicalDeviceProperies.

Change-Id: Id163482e2966f912ad3ba34d0fc7e2627dc1bfa8