vkreplay: Fix reset frame number when loopStartFrame is given.
This change sets m_frameNumber to loopStartFrame after each loop to
make the frame number counter m_frameNumber consistent with the frame
which startingPacket belongs to.

Otherwise the vkreplay will replay frames starts from loopStartFrame to
(loopStartFrame + loopEndFrame) in the second or later loop.  Because
the startingPacket is set to the first packet of frame "loopStartFrame"
but the m_frameNumber is reset to 0 which means the replayed content is
starts from the expected loopStartFrame but the internal frame counter
is starts from 0 instead of loopStartFrame.