vkreplay: Reduce CPU overhead in the main loop.
Move portability table lookup (in vkAM) out from main loop to reduce
CPU overhead in the main loop when portability table is being used.

Encapsulate all the direct call of fread/ftell/fseek with FileLike

Do alignment for the memory size needed in vkAM call to make sure it
allocates memory with the expected size during replay.  Only works when
portablility table is besing used and the alignment in vkGBMR is
different in trace platform and replay platform.