layersvt: DevSim 1.0.3 minor code reorgs
Bump version to 1.0.3

This commit is a collection of small changes that don't affect existing
functionality.  There are general code improvements, and some preparation
for new functionality in development:

1) Use JsonCpp value.type() rather than value.is*().
An old JsonCpp design decision (and tagged as will-not-change): isObject()
unfortunately returns TRUE for nullValue.  Therefore, use explicit test of
type() to ensure a value is really only an objectValue.  isArray() has same
issue.  Changing all type checks to avoid value.is*().

2) Rename ApplyOverrides() to GetValue().

3) GetValue() args changed to separate parent/name, rather than single value.

4) Minor code reorganization, comments added.

5) Simplify type checking code, since types are enforced by schema validation.

Change-Id: I7fbcc3620133efcf5b9f1ef25b4250f5f0f91b85