Add trigger to listen on a socket.
Use -tr port[-nFrames[-portNumber]]
The default port is 8100.
Then send a message on the port to trigger the capture; if
the message is a single integer, it overrides nFrames passed
on the command line.
The nFrames parameter is ignored when -portNumber is added.
It's convenient to use nc (netcat) to send the message:
  echo 200 | nc localhost 8100