Add trigger to listen on a socket.
Use -tr port[:<portnum>][,frameCount]
The default port is 8100.
Then send a message on the port to trigger the capture; if
the message is a single integer, it overrides nFrames passed
on the command line.
It's convenient to use nc (netcat) to send the message:
  echo 200 | nc localhost 8100

Only enabled on PLATFORM_LINUX.