layersvt: Devsim 1.2.1 multiple input files
Device Simulation Layer can now load multiple configuration files by specifying
a delimited list of filenames in the VK_DEVSIM_FILENAME environment variable.
The delimited list borrows the syntax of the PATH environment variable:
on Windows, separate the filenames with ';', on Linux separate with ":".

Add documentation for the multiple configuration filename to the .md file.

Some minor rearrangement of comments and debug messages.

Change-Id: If49a23fef71d35eeea2d5940bfee7ae584bd57c8