tests: DevSim 121 multiple input files
Restructure the regression test input files to exercise and make use
of the new list of configuration files feature.

The "test1" input configuration now consists of file fragments that
are all loaded at runtime to create the complete configuration.

This feature perfmits configuration files to be modular, with boilerplate data
in separate files.

Change-Id: I5d0fad652748020c3580eedfbfc677d31e6b2e01