vkreplay: Fix verification of remapped surfaces in vkCreateSwapchain.
Since pSurf points to pCreateInfo->surface, and *pSurf was just
updated to the remapped surface, this conditional was previously checking
if the surface object both WAS and WAS NOT a VK_NULL_HANDLE (which is
impossible, so this error would never get reported).

For this entrypoint, the input (trace-time) surface is not allowed to be
VK_NULL_HANDLE, so this change now verifies the surface is valid, and
reports an error if not valid.

It also verifies that the remapped surface is not VK_NULL_HANDLE; and
reports an error as needed.