vkreplay: Fix a corner case in replaying vkAM
Corner case:
A dstBuffer's memory is allocated too early and at the same time there's
a srcBuffer with the same buffer handle being created and used in
vkCmdCopyBuffer before the real dstBuffer being created and bind to the
dstBuffer's memory.

Sequence without portability table being enabled:
1. memory#dstBuffer is allocated
2. srcBuffer is created with handle#1
3. memory#srcBuffer is allocated
4. handle#1 is bind to memory#srcBuffer
5. handle#1 is used in vkCmdCopyBuffer as srcBuffer
6. handle#1 is destroyed
7. dstBuffer is created with handle#1
8. handle#1 is bind to memory#dstBuffer
9. handle#1 is used in vkCmdCopyBuffer as dstBuffer

The portability table feature will change the sequence to: (It will
create buffer before calling vkAllocateMemory if the buffer binds to the
memory is not created before allocation.)
1. dstBuffer is created with handle#1
2. memory#dstBuffer is allocated
3. srcBuffer is NOT created with handle#1, and vkreplay will use
   handle#1 directly for srcBuffer because vkreplay does not know the
   srcBuffer is different from the dstBuffer though they have the same
   buffer handle.
4. memory#srcBuffer is allocated
5. handle#1 is bind to memory#srcBuffer
6. handle#1 is used in vkCmdCopyBuffer as srcBuffer (Error happens here!
   because handle#1 was created with VK_BUFFER_USAGE_TRANSFER_DST_BIT)

This is fixed by destroy buffer/image created in
manually_replay_vkAllocateMemory() and do a clean up in object mappings.