vkreplay: GH446, Auto reorient to fit surface extent
This change is comparing content's orientation with surface's
orientation, if they are not match, vkreplay will rotate the content 90
degrees to make the content shown in expected orientation.

It is implemented by setting VK_SURFACE_TRANSFORM_ROTATE_90_BIT_KHR to
pCreateInfo->preTransform in vkCreateSwapchainKHR() when the content's
orientation does not match surface's orientation retrieved via
vkGetPhysicalDeviceSurfaceCapabilitiesKHR().