vktrace: Change use of ROUNDUP_TO_4 so mem past orig src not accessed
Whenever ROUNDUP_TO_4 is used, need to make sure that it is used to
round up the size of packet elements, but is not used as the size of
the data to be copied from the caller's data to the packet. If the
rounded up size is used for the copy, the copy could potentially
access memory past the end of a buffer/string that may not be legal
to access.

This pull request addresses this problem.

Change-Id: If7b9160611509900c9eca0c34522bd3768cd8477