vkreplay: synchronize window resizes on Linux
vkreplay currently uses a 0.05 second delay after
sending requests to the X server to resize and map
its window hoping that this will ensure it has been
processed before proceeding.

There is not guaranteed, however. Instead, it should
wait to receive a map notify event corresponding to
the request, pausing the replay meanwhile.