vkreplay: Fix error when vkAllocateMemory is skipped
The problem is vkAllocateMemory will NOT do memory allocation if a
memory is not being bound to any image or buffer.  But the memory may
be used in vkMapMemory, vkUnmapMemory, vkFreeMemory,
vkFlushMappedMemoryRanges, vkInvalidateMappedMemoryRanges which will
cause error when calling
m_objMapper.m_devicememorys.find(NOT_ALLOCATED_MEMORY)->second without
checking the NOT_ALLOCATED_MEMORY exists in m_objMapper.m_devicememorys
or not.

This fix is to skip vkMapMemory, vkUnmapMemory, vkFreeMemory,
vkFlushMappedMemoryRanges, and vkInvalidateMappedMemoryRanges when such
memory is besing used.