vkreplay: Log requested layers on failure
Log the requested extensions from the VkInstanceCreateInfo struct when
vkCreateInstance fails with VK_ERROR_LAYER_NOT_PRESENT.

No need to do the same to vkCreateDevice because enabledLayerCount
and ppEnabledLayerNames in vkDeviceCreateInfo are deprecated and
ignored.

This change also moves the restore of saved layers and extensions after
the checking of VkResult returned by vkCreateInstance to make sure the
layers/extensions being printed are the ones being used in the real
call to vkCreateInstance.