vkreplay: Re-implement cross platform replay
Allow traces created on Windows to be replayed on Linux, and vise-versa.
Cross platform replay probably requires that GPUs be from the same vendor
and that the driver versions be close.

Tested by tracing a simple program (cube) on Linux and replaying on
Windows using nVidia GPUs, and vise-versa.  This worked, but many
configurations will not work because of inherent differences in
GPUs and driver implementations.

This used to work many commits ago, got broken sometime in the past,
and it has been fixed to work again.