vktrace: Fix env clean up issue in AndroidSetEnv
The existing AndroidSetEnv() is not able to clean up an env when the val
parameter of AndroidSetEnv() is set to "" which makes len_val to be 0.

This change fixes this issue by always add quotes around val parameter.

e.g. For a call like AndroidSetEnv("key", "") trying to clean up "key"
environment variable, it will execute command 'setprop key ""' which
works as expected instead of 'setprop key'.