vktrace: Fix inconsistent use of macro in trim
This change makes the use of TRIM_USE_ORDERED_IMAGE_CREATION
consistent in vktrace.

Previously it was used in two ways:
1. #if TRIM_USE_ORDERED_IMAGE_CREATION
2. #if defined(TRIM_USE_ORDERED_IMAGE_CREATION)

Now it is used in a consistent way:
1. #if TRIM_USE_ORDERED_IMAGE_CREATION

Because TRIM_USE_ORDERED_IMAGE_CREATION is defined as 1 by default. If
the allowed value of TRIM_USE_ORDERED_IMAGE_CREATION is either 1 or 0,
then only "#if TRIM_USE_ORDERED_IMAGE_CREATION" will work as expected
when TRIM_USE_ORDERED_IMAGE_CREATION is defined as 0.