vkreplay: terminate early if xcb_connect fails
If vkreplay fails to establish a connection to the X server for whatever reason it will currently crash with a segmentation fault without providing the user with any indication as to the cause. With this change, it will check immediately whether the connection was established successfully and if not, terminate with a more informative error message.