vktrace: assign dispatch table for created objects
The vktrace layer can sometimes create new dispatchable
vulkan objects, specifically VkCommandBuffers and VkQueues.
As per the loader's documentation, if this is done within a
layer it is the layer's responsibility to assign the
dispatch table for the object as this is normally done in
the top level trampoline function which is bypassed in this
case.

However, the vktrace layer currently does not do so. This can
result in problems if it is used on top of other layers. For
example, if the validation layers are enabled and an application
is traced with the trim option it causes a segmentation fault.

Hence, this change ensures the dispatch table is correctly
assigned for these objects.