vkreplay: back out previous commit
The prior commit, ad30962a4b4dad044adfce38643f14ff4a797aad, caused
a failure in LunarG's internal CI tests on Windows.

The failure was in a replay/trace/replay of Dota2, and in a replay
of an older trace of vkQuake.

The errors introduced by the previous commit are:

   vkreplay error: Return value VK_ERROR_OUT_OF_HOST_MEMORY from API call (vkQueuePresentKHR) does not match return value from trace file VK_SUCCESS.
   vkreplay error: Failed to replay packet_id 165, with global_packet_index 2686.
   vkreplay error: Return value VK_ERROR_VALIDATION_FAILED_EXT from API call (vkAcquireNextImageKHR) does not match return value from trace file VK_SUCCESS.
   vkreplay error: Failed to replay packet_id 164, with global_packet_index 2689.
   vkreplay error: API call vkQueueSubmit returned VK_ERROR_DEVICE_LOST. vkreplay cannot continue, exiting.

Backing out this commit for now until we can isolate the issue.

Change-Id: Ib8d45807e5f84dd02338bac1c522653a54458317