vkreplay: default to x11 backend if XDG_SESSION_TYPE is unset
If no display backend is specified via the "-ds" option when running
vkreplay and the XDG_SESSION_TYPE environment variable is not set
it will trigger a segmentation fault during initialization with no
other output. This would be the case, for example, if the user is not
running a full desktop environment or is connected over SSH.

With this change, it will instead default to using the x11 backend.