vktracedump: Correct html format in full dump
This change removes GlobalPacketIndex info from full API dump in html
format to make the html format shows correctly in browser.

This change does not include GlobalPacketIndex in the html output
because that needs to chagne api dump layer's code to support
GlobalPacketIndex which only exists in a trace file and it may be
better to NOT do the change in the code of api dump layer.