vkreplay: search full portability table
Previously used to search the portability table starting
with the item after the the last item that was searched for,
and ending with the last item in the table.  This was a problem
for looped replays - the search start position is not reset at
the start of each loop, so searches would fail on all loops
after the first.

Changed code to search all entries in the table, starting
the search with the item after the  last item that was searched for.

Change-Id: Iebaaeec568f2a3d80a73c3d8fda168b0c76decf2