scripts: Modify scripts for 102 header update
The addition of the `VK_EXT_metal_surface` extension required the
addition of "metal" to the platform dictionary. The addition of the
`VK_NVX_image_view_handle` extension introduced the
`vkGetImageViewHandleNVX` function which has a return type of
`uint32_t`. This return type was not previously defined in
`layer_factory_generator.py`'s return map.

Modified:
- `scripts/common_codegen.py`
- `scripts/layer_factory_generator.py`

Change-Id: Id8761f4a109b3a0cf05e76c643bc9f1bb16a1047