vkreplay: don't allow 32-bit trace replay on 64-bit replayer
Code was already in place to enforce this, but the recently
added code to pre-read portability packets crashes when
cross-playing, so the code to enforce it was not executed.
Moved the code to check for 32/64 to earlier in the replay.

Also added enforcing of not allowing replay if the Endianess of
the trace file and replay system do not match.

Change-Id: I018a806ac9cea68ba35c572e4a4177170cf4a73b