screenshot: Use device to access correct dispatch tbl
Fixes problem with gfxrecon replay of a Dawn Of War 3 trace.

The dispatch table being used was being keyed off of the queue
handle passed in to vkQueuePresentKHR. It would work because
the queue handle was the same as the device handle for a previously
destroyed device.

Changed code to figure the device associated with the passed in
queue handle and to then use that device's displatch table.

Change-Id: Id5be18f67c48a785623c266ed884af742c378aeb