android: interactive pauses, save logs
Android tests are showing unreliable behavior when replaying
many vktrace traces; it seems the device doesn't always recognize
when an action has started.

I've added a couple of pauses as an experiment to try to get
the test to behave more like it were being driven by a human.
This seems to make at least some of the traces more reliable.

In addition, Android log files from the execution are now
saved for later examination in case of failure (or to ensure
that the device behaved as expected).