screenshot: Delete reused queue and cmdbuf keys from unordered maps
If a queue or cmdbuf handle is reused, the attempt to add an
entry to an unordered map using the reused queue or cmdbuf
handle as a key will fail. Added coded to remove the map entry
that is using the reused queue/cmdbuf key before adding the new
entry to the unordered map.

The following sequence caused this error:
   vkCreateDevice() returns device D1
   vkGetDeviceQueue(D1) returns queue Q1
   vkDestroyDevice(D1)
   vkCreateDevice() returns device D2
   vkGetDeviceQueue(D2) returns queue Q1
   vkQueuePresent(Q1)

The second call to vkGetDeviceQueue would not add Q1 to an unordered map
(leaving the previous entry in place), resulting in the vkQueuePresent
attempting to use the dispatch table from destroyed device D1.

An alternative fix would have been to change the behavior of vkDestroyDevice
to delete queue/cmdbuf entries that share the dispatch table with the device
that is being destroyed. Such a list wasn't readily available, so used the
above scheme instead.