vktrace: Better errors on read failures
Occasional errors occur while reading that are difficult to track
down because insufficient error information is given.  These
changes add more error information to existing error paths, including
an error message on a potential short read (which silently failed before).

Also added a new verbosity level "max" that is identical to "full" on
Release builds, and "debug" on Debug builds.

Cleaned up some of the descriptions in vktrace.md; there were several
unquoted uses of "_" that have to be quoted in markdown.

Change-Id: I9fca65846f7e54f829c43f7e2b5c9e7e1b42449d