apidump: Timestamps in dump output
apidump will now show the time a function was called since start
in microseconds. This is useful to determine which functions are
taking significant time to execute.

added the VK_APIDUMP_TIMESTAMP environmental variable and the
lunarg_api_dump.show_timestamp setting to the api_dump settings file

changed files:
	layersvt/api_dump.h
	layersvt/api_dump_layer.md
	layersvt/vk_layer_settings.txt
	scripts/api_dump_generator.py
	vkconfig/layer_info.json

Change-Id: Ie223747d947c803617018034b7c5f66d3a243d9b