tests: run all trace tests even after failure
Currently, the script will exit after the first test failures.
It's helpful to continue to run the rest of the tests after a
failure; this can help determine whether a failure is
trace-specific or systemic, and ensures that a tester has a full
set of pass/fail information after a test run.

A switch is provided to restore the original behavior, if this
is desired.

Change-Id: Ie709a1123394d405f558ebd7656cd6f0b41650a1