vkvia: update search for installed sdk packages
Update the Vulkan SDK package search for 'vulkan-sdk' on only the
ubuntu distro.

The 'lunarg-vulkan-sdk' has long since been changedcc to just
'vulkan-sdk'.  The 'vulkan-sdk' package is only provided for
Ubuntu so there is no need to check for this package on any
other distros.

Arch Linux used to have a lunarg-vulkan-sdk package but has also long
been marked as oboslete and they have since provided packages for each
repository in the SDK.