vkreplay: Change algorithm when searching for queue family index
Changed algorithm when searching for a replay device queue family index
when in portability mode. The search starts with the queue family index
in the trace file so as to increase the probablily that the queue
family index selected will work.

Also changed the name of the method getQueueFamilyIdx to getReplayQueueFamilyIdx,
changed it to have no return value, and removed one arg. This simplified code
that calls getReplayQueueFamilyIdx.

Change-Id: Ic09ac4bff3bfff7330c229c3279ac3cbdb65e87b