api_dump: keep lock held between printing head and body
Previously, api_dump would release the output lock when finished printing
the head of a call and reaquiring it when printing the body. This enabled
multi threaded applications to interleave output resulting in garbled output.
This commit fixed that by keeping the output lock held for the duration of
the call, at the cost of serializing any calls that are made to vulkan.

This commit also switched api_dump to std::mutex & std::thread::id instead of using
vk_loader_platform.h for the threading primitives.

Changes:
	modified:   layersvt/api_dump.h
	modified:   scripts/api_dump_generator.py

Change-Id: Ia4b0cabc839c93798eeeb811d6cb322b5b197194